colocate useAttachmentLoader and getFilepathForLoadable
