+++ src/js/express-xr/src/SceneManagerXR.js
@@ -41,7 +41,8 @@ const GUI = require('./gui/GUI')
 
 const { controllerObjectSettings, cameraObjectSettings } = require('./utils/xrObjectSettings')
 const { getIntersections, boneIntersect, intersectObjects, constraintObjectRotation, setControllerData } = require('./utils/xrControllerFuncs')
-const { findParent, moveObject, rotateObject, createHideArray, getFilepathForLoadable, updateObjectHighlight } = require('./utils/xrHelperFuncs')
+const { findParent, moveObject, rotateObject, createHideArray, updateObjectHighlight } = require('./utils/xrHelperFuncs')
+const { useAttachmentLoader, getFilepathForLoadable } = require('./hooks/useAttachmentLoader')
 const applyDeviceQuaternion = require('../../shot-generator/apply-device-quaternion')
 require('./lib/VRController')
 
@@ -59,91 +60,6 @@ new THREE.AudioLoader().load('data/snd/vr-select.ogg', () => {})
 new THREE.AudioLoader().load('data/snd/vr-welcome.ogg', () => {})
 new THREE.AudioLoader().load('data/snd/vr-beam2.mp3', () => {})
   // new THREE.AudioLoader().load('data/snd/vr-atmosphere.mp3', () => {})
-
-const useAttachmentLoader = ({ sceneObjects, world }) => {
-  // TODO why do PENDING and SUCCESS get dispatched twice?
-  const [attachments, dispatch] = useReducer((state, action) => {
-    switch (action.type) {
-      case 'PENDING':
-        // ignore if already exists
-        return (state[action.payload.id])
-          ? state
-          : {
-              ...state,
-              [action.payload.id]: { status: 'NotAsked' }
-            }
-      case 'LOAD':
-        // ignore if already loading
-        return (state[action.payload.id].loading)
-          ? state
-          : {
-              ...state,
-              [action.payload.id]: { status: 'Loading', progress: undefined }
-            }
-      case 'PROGRESS':
-        return {
-          ...state,
-          [action.payload.id]: {
-            ...[action.payload.id],
-            progress: {
-              loaded: action.payload.progress.loaded,
-              total: action.payload.progress.total,
-              percent: Math.floor(action.payload.progress.loaded/action.payload.progress.total) * 100
-            }
-          }
-        }
-      case 'SUCCESS':
-        return {
-          ...state,
-          [action.payload.id]: { status: 'Success', value: action.payload.value }
-        }
-      case 'ERROR':
-        return {
-          ...state,
-          [action.payload.id]: { status: 'Error', error: action.payload.error }
-        }
-      default:
-        return state
-    }
-  }, {})
-
-  useMemo(() => {
-    let loadables = Object.values(sceneObjects)
-      // has a value for model
-      .filter(o => o.model != null)
-      // has not loaded yet
-      .filter(o => o.loaded !== true)
-      // is not a box
-      .filter(o => !(o.type === 'object' && o.model === 'box'))
-
-    world.environment.file && loadables.push(
-      { type: 'environment', model: world.environment.file }
-    )
-
-    loadables.push(controllerObjectSettings)
-    loadables.push(cameraObjectSettings)
-
-    loadables.forEach(o =>
-      dispatch({ type: 'PENDING', payload: { id: getFilepathForLoadable({ type: o.type, model: o.model }) } })
-    )
-  }, [sceneObjects])
-
-  useMemo(() => {
-    Object.entries(attachments)
-      .filter(([k, v]) => v.status === 'NotAsked')
-      .forEach(([k, v]) => {
-        gltfLoader.load(
-          k,
-          value => dispatch({ type: 'SUCCESS', payload: { id: k, value } }),
-          progress => dispatch({ type: 'PROGRESS', payload: { id: k, progress } }),
-          error => dispatch({ type: 'ERROR', payload: { id: k, error } })
-        )
-        dispatch({ type: 'LOAD', payload: { id: k } })
-      })
-  }, [attachments])
-
-  return attachments
-}
 
 const useVrControllers = ({ onSelectStart, onSelectEnd, onGripDown, onGripUp, onAxisChanged }) => {
   const [controllers, setControllers] = useState([])
